pickfans/requirements.txt
asgiref==3.2.3
aspy.yaml==1.3.0
attrs==19.3.0
CacheControl==0.12.6
cachy==0.3.0
certifi==2019.11.28
cfgv==2.0.1
chardet==3.0.4
cleo==0.7.6
clikit==0.4.1
Django==3.0.2
django-environ==0.4.5
djangorestframework==3.11.0
elasticsearch==7.1.0
html5lib==1.0.1
identify==1.4.9
idna==2.8
importlib-metadata==0.23
importlib-resources==1.0.2
jsonschema==3.2.0
keyring==19.3.0
lockfile==0.12.2
more-itertools==8.0.2
msgpack==0.6.2
mysqlclient==2.0.1
nodeenv==1.3.3
pastel==0.1.1
pexpect==4.7.0
pkginfo==1.5.0.1
poetry==1.0.0
pre-commit==1.20.0
ptyprocess==0.6.0
pylev==1.3.0
pyparsing==2.4.6
pyrsistent==0.14.11
pytz==2019.3
PyYAML==5.2
requests==2.22.0
requests-toolbelt==0.8.0
shellingham==1.3.1
six==1.13.0
sqlparse==0.3.0
toml==0.10.0
tomlkit==0.5.8
urllib3==1.25.7
uWSGI==2.0.19.1
virtualenv==16.7.9
webencodings==0.5.1
zipp==0.6.0
